style fixes
