+++ core/src/main/scala/com/spotify/noether/ConfusionMatrixAggregator.scala
