+++ core/src/main/scala/com/spotify/noether/LogLossAggregator.scala
@@ -19,8 +19,10 @@ package com.spotify.noether
 
 import com.twitter.algebird.{Aggregator, Semigroup}
 
-case object LogLossAggregator extends Aggregator[Prediction[Int, List[Double]], (Double, Long), Double] {
-  def prepare(input: Prediction[Int, List[Double]]): (Double, Long) = (math.log(input.predicted(input.actual)), 1L)
+case object LogLossAggregator
+  extends Aggregator[Prediction[Int, List[Double]], (Double, Long), Double] {
+  def prepare(input: Prediction[Int, List[Double]]): (Double, Long) =
+    (math.log(input.predicted(input.actual)), 1L)
   def semigroup: Semigroup[(Double, Long)] = implicitly[Semigroup[(Double, Long)]]
   def present(score: (Double, Long)): Double = -1*(score._1/score._2)
 }
