+++ core/src/main/scala/com/spotify/noether/MulticlassConfusionMatrixAggregator.scala
