+++ core/src/main/scala/com/spotify/noether/Prediction.scala
