+++ core/src/test/scala/com/spotify/noether/ErrorRateAggregatorTest.scala
@@ -19,16 +19,16 @@ package com.spotify.noether
 
 import org.scalactic.TolerantNumerics
 
-class LogLossTest extends AggregatorTest {
+class ErrorRateAggregatorTest extends AggregatorTest {
   private implicit val doubleEq = TolerantNumerics.tolerantDoubleEquality(0.1)
   private val classes = 10
-  private def s(idx: Int, score: Double): List[Double] =
-    0.until(classes).map(i => if(i == idx) score else 0.0).toList
+  private def s(idx: Int): List[Double] =
+    0.until(classes).map(i => if(i == idx) 1.0 else 0.0).toList
 
   it should "return correct scores" in {
-      val data = List((s(0, 0.8), 0), (s(1, 0.6), 1), (s(2, 0.7), 2))
+      val data = List((s(1), 1), (s(3), 1), (s(5), 5), (s(2), 3), (s(0), 0), (s(8), 1))
         .map{case(scores, label) => Prediction(label, scores)}
 
-      assert(run(LogLossAggregator)(data) === 0.363548039673)
+      assert(run(ErrorRateAggregator)(data) === 0.5)
     }
 }
