+++ core/src/test/scala/com/spotify/noether/LogLossAggregatorTest.scala
@@ -19,20 +19,16 @@ package com.spotify.noether
 
 import org.scalactic.TolerantNumerics
 
-class ConfusionMatrixTest extends AggregatorTest {
+class LogLossAggregatorTest extends AggregatorTest {
   private implicit val doubleEq = TolerantNumerics.tolerantDoubleEquality(0.1)
+  private val classes = 10
+  private def s(idx: Int, score: Double): List[Double] =
+    0.until(classes).map(i => if(i == idx) score else 0.0).toList
 
   it should "return correct scores" in {
-    val data =
-      List((0.1, false), (0.1, true), (0.4, false), (0.6, false), (0.6, true), (0.6, true), (0.8, true))
-        .map{case(s, pred) => Prediction(pred, s)}
+      val data = List((s(0, 0.8), 0), (s(1, 0.6), 1), (s(2, 0.7), 2))
+        .map{case(scores, label) => Prediction(label, scores)}
 
-    val matrix = run(ConfusionMatrixAggregator())(data)
-
-    assert(matrix.tp === 3L)
-    assert(matrix.fp === 1L)
-    assert(matrix.fn === 1L)
-    assert(matrix.tn === 2L)
-  }
+      assert(run(LogLossAggregator)(data) === 0.363548039673)
+    }
 }
-
